fix:lowcase수정
