+++ client/src/components/Community/CommunityBoard.jsx
@@ -26,56 +26,77 @@ const getCategoryName = categoryNumber => {
 // 커뮤니티 게시판 컴포넌트
 const CommunityBoard = () => {
   const navigate = useNavigate();
-  const loginUser = useRecoilValue(loginState);
 
+  // 로그인이 안되면 로그인 페이지로 이동
+  const loginUser = useRecoilValue(loginState);
   if (loginUser?.id === '' && loginUser?.email === '') navigate('/login');
 
+  // Recoil 상태로부터 검색어 가져오기
   const searchKeyword = useRecoilValue(CommunitysearchKeywordState);
+  // 게시물 상태 관리
   const [allPosts, setAllPosts] = useState([]);
+  // 선택된 카테고리 상태 관리
   const [selectedCategory, setSelectedCategory] = useState('all');
+
+  // 현재 페이지와 페이지 당 표시할 게시물 수 설정
   const [currentPage, setCurrentPage] = useState(1);
   const postsPerPage = 10;
 
   useEffect(() => {
-    const getAllPosts = async () => {
+    // API 호출하여 게시물 데이터 가져오기
+    const fetchAllPosts = async () => {
       try {
+        // const response = await fetch('http://localhost:8001/community');
         const response = await axios.get('http://heallenges.cafe24app.com/community');
-        // Ensure that the server response has the expected structure
-        const data = response.data.data || [];
-        setAllPosts(data);
+        const data = await response.json();
+        setAllPosts(data.data || []);
       } catch (error) {
         console.error('데이터 가져오기 오류:', error);
       }
     };
 
-    getAllPosts();
+    // 컴포넌트가 마운트될 때 한 번 실행
+    fetchAllPosts();
   }, []);
 
+  // 검색된 게시물 필터링
+  // const searchallPosts = allPosts.filter(data => data?.title.toLowerCase().includes(searchKeyword.toLowerCase()));
+  const searchallPosts = allPosts.filter(data => {
+    const title = data?.title;
+    return typeof title === 'string' && title.includes(searchKeyword);
+  });
+
+  // 카테고리 변경 핸들러
   const handleCategoryChange = category => {
+    // 선택된 카테고리가 변경되면 현재 페이지를 1로 초기화
     setCurrentPage(1);
     setSelectedCategory(category === 'all' ? 'all' : parseInt(category));
   };
 
-  const filteredBoardList = allPosts;
-  // .filter(post => selectedCategory === 'all' || post.category === selectedCategory)
-  // .filter(data => {
-  //   const title = data?.title;
-  //   return typeof title === 'string' && title.toLowerCase().includes(searchKeyword.toLowerCase());
-  // });
+  // 필터된 게시물 목록
+  const filteredBoardList = searchallPosts.filter(
+    post => selectedCategory === 'all' || post.category === selectedCategory,
+  );
 
+  // 현재 페이지에 표시할 게시물의 범위 계산
   const indexOfLastPost = currentPage * postsPerPage;
   const indexOfFirstPost = Math.max(0, indexOfLastPost - postsPerPage);
   const currentPosts = filteredBoardList.slice(indexOfFirstPost, indexOfLastPost);
 
+  // 페이지 변경 이벤트 핸들러
   const handlePageChange = pageNumber => {
     const newPageNumber = Math.min(Math.max(1, pageNumber), Math.ceil(filteredBoardList.length / postsPerPage));
     setCurrentPage(newPageNumber);
   };
 
+  // JSX 반환
   return (
     <div className="card mb-4">
       <div className="mt-3">
+        {/* 게시판 제목 */}
         <h3>&nbsp;게시판 목록</h3>
+
+        {/* 카테고리 선택 버튼들 */}
         <div className="mb-2 d-flex justify-content-end">
           <button className="btn btn-outline-secondary me-1" onClick={() => handleCategoryChange('all')}>
             전체
@@ -92,6 +113,7 @@ const CommunityBoard = () => {
         </div>
       </div>
 
+      {/* 게시물 목록을 표시하는 테이블 */}
       <table className="table">
         <thead>
           <tr>
@@ -105,11 +127,14 @@ const CommunityBoard = () => {
           </tr>
         </thead>
         <tbody>
+          {/* 현재 페이지에 표시할 게시물들 매핑 */}
           {currentPosts.map((post, index) => (
             <tr key={post.id}>
+              {/* 게시물 정보 표시 */}
               <th scope="row">{indexOfFirstPost + index + 1}</th>
               <td>{getCategoryName(post.category)}</td>
               <td>
+                {/* 게시물 제목에 링크 추가 */}
                 <Link to={`/community/${post.id}`}>{post.title}</Link>
               </td>
               <td>{post.nickname}</td>
@@ -121,6 +146,7 @@ const CommunityBoard = () => {
         </tbody>
       </table>
 
+      {/* 페이지네이션 */}
       <nav className="d-flex justify-content-center">
         <Pagination
           activePage={currentPage}
